Move mergeAggregationsResposes to utils
