+++ lib/resources.js
@@ -18,7 +18,14 @@ const { parseParams } = require('../lib/util')
 const { AGGREGATIONS_SPEC } = require('./elasticsearch/config')
 
 const errors = require('./errors')
-const { esRangeValue, parseSearchParams, nyplSourceAndId, itemsByFilter } = require('./utils/resource-helpers')
+const {
+  esRangeValue,
+  parseSearchParams,
+  nyplSourceAndId,
+  itemsByFilter,
+  mergeAggregationsResponses
+} = require('./utils/resource-helpers')
+
 const {
   bodyForFindByUri,
   addInnerHits,
@@ -206,33 +213,6 @@ module.exports = function (app, _private = null) {
         }
         throw new IndexSearchError(`Error processing search: ${e.message || e}`)
       })
-  }
-
-  /**
-  * Given an array of ES aggregations responses (such as that returned from msearch)
-  **/
-  const mergeAggregationsResponses = (responses) => {
-    // Filter out errored responses:
-    responses = responses.filter((resp) => resp.aggregations)
-    if (responses.length === 0) {
-      return {}
-    }
-    return {
-      // Use `hits` of last element, somewhat arbitrarily:
-      hits: responses[responses.length - 1].hits,
-      aggregations: responses
-        .reduce((allAggs, resp) => {
-          const respAggs = Object.entries(resp.aggregations)
-            // Build hash of response aggs, squashing _nested aggs:
-            .reduce((a, [field, _a]) => {
-              // If it's nested, it will be in our special '_nested' prop:
-              a[field] = _a._nested || _a
-              return a
-            }, {})
-          // Add response aggs to combined aggs:
-          return Object.assign(allAggs, respAggs)
-        }, {})
-    }
   }
 
   // Get all aggregations:
