+++ lib/utils/resource-helpers.js
@@ -115,9 +115,37 @@ function itemsByFilter (identifierValues, app) {
     })
 }
 
+/**
+* Given an array of ES aggregations responses (such as that returned from msearch)
+**/
+const mergeAggregationsResponses = (responses) => {
+  // Filter out errored responses:
+  responses = responses.filter((resp) => resp.aggregations)
+  if (responses.length === 0) {
+    return {}
+  }
+  return {
+    // Use `hits` of last element, somewhat arbitrarily:
+    hits: responses[responses.length - 1].hits,
+    aggregations: responses
+      .reduce((allAggs, resp) => {
+        const respAggs = Object.entries(resp.aggregations)
+          // Build hash of response aggs, squashing _nested aggs:
+          .reduce((a, [field, _a]) => {
+            // If it's nested, it will be in our special '_nested' prop:
+            a[field] = _a._nested || _a
+            return a
+          }, {})
+        // Add response aggs to combined aggs:
+        return Object.assign(allAggs, respAggs)
+      }, {})
+  }
+}
+
 module.exports = {
   esRangeValue,
   parseSearchParams,
   nyplSourceAndId,
-  itemsByFilter
+  itemsByFilter,
+  mergeAggregationsResponses
 }
